Update README.md
I am glad that I complete my front-end web page using HTML and CSS.
It took me about 8 hours to complete it. 
I am proud of myself for being able to create a web page from scratch. 
It was a challenging task that required a lot of time and effort, but it was worth it in the end.
Front-end development is an essential part of web development that focuses on creating the user interface and user experience of a website. 
HTML and CSS are two of the most fundamental technologies used in front-end development. 
HTML provides the structure of the web page, while CSS is used for styling and layout.
I am satisfied with how my web page looks. 
It's always satisfying to see the results of your hard work. 
I hope that others will find my web page useful and informative. 
If you have any questions feel free to ask to contact me.